moved cimcore tests to its own test suite.
there are currently no tests to run as the test samples need migration to CIMPL6 grammar and also aligned with current SHR models.

This change is to just make the import-test.js file cleaner.
